Reuse function docs for matching parameter sets.
This causes the documentation for a given parameter set to be used
on all matching overloads. The error message is preserved as an
informational message.

The rationale for this is that this documentation should be
applicable to these matching overloads.

It is not currently clear if there is a need for different
documentation on these matching overloads. If that is the case,
the parameters need to support annotating an optional type
signature and using that in the overload lookup.